fix: stale refill rate after set_max_capacity, unbound var in refund, type annotation, and NaN/inf validation
- Recalculate _rate_per_sec in set_max_capacity() and get_max_capacity()
  across all three bucket types so the refill rate stays in sync with
  the cap (Bug 1, high severity)
- Move updated_capacities assignment out of for-loop in Redis
  refund_capacity() to prevent UnboundLocalError on empty
  prerefund_capacities (Bug 2)
- Fix tuple[str, float] → tuple[str, int] type annotation in all four
  backend refund_capacity() methods (Bug 3)
- Validate max_capacity is finite and positive (math.isfinite + > 0)
  in all set_max_capacity() write paths and get_max_capacity() Redis
  read paths to reject NaN/inf which would bypass rate limiting

+++ token_throttle/_limiter_backends/_memory/_sync_backend.py
@@ -270,7 +270,7 @@ class SyncMemoryBackend(SyncRateLimiterBackend):
             )
 
             # Apply refund amounts to current capacity
-            updated_capacities_: dict[tuple[str, float], float] = dict(
+            updated_capacities_: dict[tuple[str, int], float] = dict(
                 prerefund_capacities,
             )
             for cap_metric, per_seconds in prerefund_capacities:
